requirements.txt
openai>=1.0.0
streamlit>=1.24.0
python-dotenv>=0.19.0
Pillow>=8.0.0
